Add support for Nextar tracking commands
You can now specify the mount type so that we can
switch between tracking on and off.  Note that the
mount type and tracking mode can not be changed other than via the CLI.

Meaning, you can't switch between AltAz and EQ-North, but you can turn
on and off tracking.  This is because ASCOM/Alpaca does not support the
concept of tracking modes which AFAIK only really exists with AltAz
mounts that optionally support a wedge... just tracking.

Fixes: #15